+++ src/controllers/categories.controller.ts
@@ -26,9 +26,3 @@ export const getCategories = (_req: Request, res: Response): void => {
     msg: 'get categories'
   })
 }
-
-export const getCategoryMovies = (_req: Request, res: Response): void => {
-  res.json({
-    msg: 'get movies of category :id'
-  })
-}