+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Added
 
-- todo: Get movies route with query functionalities.
+- Get movies route with query functionalities.
 - todo: Connected verifyToken to some routes that require authentication.
 - todo: clean and work on the other functionalities.
 - todo: Write simple readme with how to setup the app. It has placeholders for where it would be deployed.